Move RelationshipKeySupport util class
