+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/util/RelationshipKeySupport.java
@@ -25,7 +25,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.tracker.preheat;
+package org.hisp.dhis.tracker.util;
 
 import static org.apache.commons.lang3.StringUtils.trimToEmpty;
 
@@ -40,7 +40,7 @@ import org.hisp.dhis.tracker.domain.Relationship;
 import org.hisp.dhis.tracker.domain.RelationshipItem;
 
 @NoArgsConstructor( access = AccessLevel.PRIVATE )
-public class RelationshipPreheatKeySupport
+public class RelationshipKeySupport
 {
 
     public static boolean hasRelationshipKey( Relationship relationship )
